Use "short" type names in casts

+++ src/sad_spirit/pg_wrapper/converters/datetime/IntervalConverter.php
@@ -123,7 +123,7 @@ class IntervalConverter extends BaseConverter
             $interval->s = (int)$parts[2];
         } else {
             $interval->s = (int)substr($parts[2], 0, $pos);
-            $interval->f = (double)substr($parts[2], $pos);
+            $interval->f = (float)substr($parts[2], $pos);
         }
     }
 
@@ -160,7 +160,7 @@ class IntervalConverter extends BaseConverter
                     } else {
                         $interval->s = (int)substr($tokenValue, 0, $pos);
                         $interval->f = ('-' === $tokenValue[0] ? -1 : 1)
-                                       * (double)substr($tokenValue, $pos);
+                                       * (float)substr($tokenValue, $pos);
                     }
 
                     $keys = [$intervalKey];
@@ -327,7 +327,7 @@ class IntervalConverter extends BaseConverter
                     $interval->d = 7 * (int)substr($m['w'], 0, -1);
                 } elseif ('s' === $key && false !== ($pos = strpos($m['s'], '.'))) {
                     $interval->s = (int)substr($m['s'], 0, $pos);
-                    $interval->f = ('-' === $m['s'][0] ? -1 : 1) * (double)substr($m['s'], $pos, -1);
+                    $interval->f = ('-' === $m['s'][0] ? -1 : 1) * (float)substr($m['s'], $pos, -1);
                 } else {
                     $interval->{$key} = (int)substr($m[$key], 0, -1);
                 }
